rasg/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
scikit-learn
pydantic
pdfplumber
beautifulsoup4
python-multipart
requests
spacy
openai
pytest